Remove cycle  browserHostService.ts -> web.api.ts. (#198221)
* Remove cycle  browserHostService.ts -> web.api.ts.

Before this change, `IWorkbenchConstructionOptions` in web.api.ts
depended on `IWorkspaceProvider` in browserHostService.ts, which
depends on some types from web.api.ts.

While this apparently does not pose problems in VS Code's build, other
build system of downstream projects do not allow this, and it is also
overall harder to follow the code if "the API" depends on types defined
in ""the implementation".

* Fix one reference I missed
